+++ bootstrap/providers.php
@@ -2,8 +2,10 @@
 
 use App\Providers\AppServiceProvider;
 use App\Providers\FortifyServiceProvider;
+use App\Providers\ViewServiceProvider;
 
 return [
     AppServiceProvider::class,
     FortifyServiceProvider::class,
+    ViewServiceProvider::class,
 ];