+++ database/seeders/DatabaseSeeder.php
@@ -16,6 +16,7 @@ class DatabaseSeeder extends Seeder
     public function run(): void
     {
         $this->call(SystemPermissionsSeeder::class);
+        $this->call(FirstSuperadminSeeder::class);
 
         // User::factory(10)->create();